Adding `transform: scale(1)` to F&D anchor
In addtion to `backface-visibilty: hidden`, `transform: scale(1)` has
been added to fix a bug across Chrome, Safari and Firefox.

Initially, `backface-visibility` was added because of a bug in Chrome,
FF and Opera. When an item in the right column was hovered, the
content would disappear. Interestingly, `backface-visibility` had the
complete opposite effect in Safari; all of the content was hidden.
Adding `transform: scale(1)` fixes Safari and has no effect on other
browsers.

The issue stems from the use of CSS3 columns. It seems there is a bug
when a transition is used inside of a column item.