Improve sub-nav performance
The sub-nav was being cloned and then re-appended to the DOM; I
suspected this to be less performant than just adding a class to
the already existing sub-nav. I created a jsperf test to confirm:
http://jsperf.com/sticky-navigation-clone-and-append-vs-adding-class

My hope is that this update will allow the sub-nav to perform better
on mobile.