Drop flexbox from tours layout
To fix an earlier layout bug in the tours component Flexbox was added to make the items justify left when there was an uneven number of tour items. But we discovered that the layout broke in Safari, even though Flexbox is supported. Quickest fix was to drop flexbox and use good old fashioned floats.