Hide sub nav when window is scrolled too fast on mobile
There's an issue on mobile devices when the window is scrolled too
fast to the top, the sub nav doesn't have time to unstick itself in
time and appears above the header and masthead. To "fix" this bug, I
just added a z-index on the masthead one higher than the content area
and added a white background to the header. This way, if the sub nav
doesn't unstick in time, at least it's hidden.