Improve performance of fixed sub-nav on mobile when scrolling
There's an issue on iOS Safari where the sub-nav is not visible and
does not stick until the scroll is complete. To fix, hardware
acceleration has been forced by adding `transform: translateZ(0);` to
the sub-nav element.

Also, decreased the height of `sub-nav__container` so that the border
does not sit below the white area of the sub-nav.