Add SVG icon sprite
This will require a larger discussion as we're currently using fonts for icons,
but this was easier and faster than trying to add stuff to icomoon (at least for me).
I also didn't want to inline the images via data URIs; it's easy to add but not
easy to update, messy and not cached by the browser (could be wrong). I definitely
think SVG is the way to go, but let's discuss when we have time.