Remove reveal animation on div. Add to text;
This tweaks the styling for the beta opt-in/out banner. It was decided that having the banner animate open for every new page load would be a poor user experience. Animating the opacity of the text helps mitigate the "pop-in" effect from rendering on the client side a little bit.