Improve scroll to next feature for article experience
This update changes how the scroll to next feature works.

The calculations for loading articles has been improved. Instead of
using each article's height and top offset, calculations using the
document and window height are used. This allows the new article
to be loaded near the bottom of the page and it removes the need to
use `setTimeout` to get calculations.

Some updates have been made in order to animate the new article into
place. Mostly just adding classnames and using the `waitForTransition`
method.

A promise has been created for the next article AJAX call.