Update analytics.js to better track events
Category should be something specific, i.e., what is it that we're tracking? Example: article-gallery
Event should be an event, i.e., click, scroll, hover, etc.
Label should be a string and describe what is being tracked, i.e., pathname for article loaded via ajax

This will probably need to be refactored so that we're not passing an object to label and so we don't need to use condtionals.