Rename custom fonts
"Main" has been renamed "Benton Sans"; "secondary" has been renamed "Miller Daily".

These names are now representative of the each font's actual name. Ideally, the fonts should be declared by their actual names and not a keyword. "Main" and "secondary" are too ambiguous. What would happen if a third font was added? Consider the approach taken by Typekit and Google Fonts.

Anyone looking at the code now will be able to identify what font is being used where without having to know that a specific keyword matches a certain font.

This update will also benefit the webfont loader because the fonts need to be declared within the loader, so it'll help if the names are descriptive.