Improve things to do component
There were some layout and design issues with the component. There was also some side scrolling on iOS. Additionally, there was some unnecessary complexity in the CSS code which has been removed. Lastly, the markup for the more and less buttons have been removed from the JS and placed within a handlebars template.