Update sub-nav component, experiences (ttd) link
The experiences (ttd) link is being added dynamically now and it wasn't getting the active state when the section was scrolled into. This was fixed by moving the `subscribe` and `addClientSideComponents` methods above the scroll-spy code and inside of the conditional.

Completed a TODO which loads dynamic sub nav items in via a handlebars template and allows for dynamic data (id and title).

Changed `#ttd` to `#experiences`. Just a small change but makes things a little nicer.