Fix browser history bug with map
Fixes lonelyplanet/destinations-next#455

Added a state object when using `pushState` and then used the `window.onpopstate` method to check the state.

Removed the `createHistory` module import. I'm not sure why that was being used instead of the native `window.history`. The `createHistory` module seemed to mess up the value of the state when it was updated with `pushState`. It would change it from a boolean to a hashed string (something like `{isOnMap: "xY9ds"}`) which would break functionality.