Fix issues with POI callout and refactor
On articles with large images, the height of the article was calculated incorrectly and so POI callouts at the end of the article were appearing too high on the page and out of view. This problem was solved by recalculating the article height one time on the scroll event.

Added line clamping back.

Also some refactoring:

- Namespace component
- Create template for callout content, move template for callout out of JS