Add variables for black and white
It may seem silly to create variables for true black and white because these colors are easy to remember. However, these variables will help with instances of `rgba` and the grayscale pallete, which is yet to be defined, but some grays have been created with Sass’ `lighten` or `darken` functions.

It also allows us to be consistent by keeping all colors assigned to a variable and we can check that all colors have a variable with the `ColorVariable` lint rule.