Update colors without variables
This step was to find all hex values without an assigned variable. In some instances, these colors were close to exisiting colors so those variables were used to replace the hex value. In the other cases, a new variable was created for the color.

This was also done to make the `ColorVariable` lint rule pass. We'll keep all colors assigned to a variable going forward.