Fix navigation bug in IE9
The navigation menu was only invisible not actually hidden, so in some
cases (IE9), hovering over the area that the menu occupies would trigger
it and make it visible. `display` could not be used since it doesn't
allow for the animation, so `visibility` was used.

Also, fixes the logo position in IE9.

Closes #553