Fix build issue
This highlights a bigger issue which is that the header build process desperately needs to be updated.

When the `node bin/rizzo build -c ../lib/data/header_normal.json` was added to `npm run build`, it was added after the original `node bin/rizzo build`. This was causing the `rizzo-next.js` file to only contain the constructors listed in... https://github.com/lonelyplanet/rizzo-next/blob/master/lib/data/header_normal.json.

The build process for the header was a bandaid for getting us a global nav and we need to pay down that technical debt.