Update masthead ad style to work with multiple ads
With the addition of the destinations book ad, it’s possible that there
could be up to 3 ads in the space. Flexbox layout was placed on the
masthead ad container so that the ads can stack vertically. text-align
was set to “left”. It used to be “right” because when there was one
single sponsor logo, it was placed at the bottom right of the masthead.

The “best in” logo is placed bottom left and so is the new book ad.

Left padding was added to match the right for when there are
left-aligned ads.