Add empty ad slot to masthead for new cities book ad (#641)
* Add ad template to masthead for book ad

* Delete extra whitespace from masthead

* Update masthead ad style to work with multiple ads

With the addition of the destinations book ad, it’s possible that there
could be up to 3 ads in the space. Flexbox layout was placed on the
masthead ad container so that the ads can stack vertically. text-align
was set to “left”. It used to be “right” because when there was one
single sponsor logo, it was placed at the bottom right of the masthead.

The “best in” logo is placed bottom left and so is the new book ad.

Left padding was added to match the right for when there are
left-aligned ads.

* Force sponsor logo ad to bottom right of masthead

With the addition of third ad to the masthead and the introduction
of flexbox to help with layout, the sponsor logo needs to be forced
into its original position. Previously, this was achieved by aligning
the ad content right. But this doesn't work now with three ads.

* Replace float with top margin for best in ad

The float does nothing now that flexbox is used to layout the masthead
ads. Top margin is needed because the new book ad stacks on top of the
best in ad and if they’re both visible, they need some space between
them.