fix: Add SVG icons to global footer (#693)
For newer Node/React apps that require the header and footer from Rizzo
Next, this should always include the required icons for the footer.
This is the same approach taken with the header. This means that the
entire suite of SVG icons from Rizzo Next doesn’t need to be required
for new apps.