feat: Delete .eot font embeds
.eot font files are specifically used for IE 6 - 11 and we don’t
support IE below 11, so they are no longer needed.

.woff has support pretty good support across the board; from IE 9 and
up and all of Edge, but no support in Android 4.3 and below.

.ttf files have the overall best support for legacy browsers,
specifically for Android 4.3 and below. I’m not sure what our support
for legacy Android is, so .ttf remains. However, if we feel fine serving
a non-custom font to legacy Android, then we can remove .ttf.

I would like to see if we can get a .woff2 version of our fonts. The
file sizes are much smaller on average and support is pretty good.
Ideally we would only load .woff and .woff2 files.