fix: reallow basic text search from nav
The search bar overlay was at one point working but then deactivated. My
best understanding was that it was the result of autocomplete not
working. Reimplementing without autocomplete to allow a user to make a
search query from anywhere on the site without having to drill down into
the search page first.