fix: clean dfp strings of special characters
Google ads are very particular about which special characters are
allowed and which are not. Many of our place names contain special latin
characters, commas or parentheses. Clean these before slugifying and
passing to DFP as key values. This will fix an issue with two places in
the upcoming BiT campaign that should be targeted.